requirements.txt
python-benedict==0.32.0
numpy==1.25.2
pandas==2.0.3
scikit-learn==1.2.2
tqdm==4.66.1
neurokit2==0.2.7
notebook==7.0.7
ipywidgets==8.0.4
